# backend/src/main/resources/swagger.yml
openapi: 3.0.3
info:
  title: Aluguel de Carros API
  description: |
    API REST do serviço de aluguel de carros (Micronaut).

    **Autenticação:** a maioria das rotas exige header `Authorization: Bearer <JWT>`.

    **Rotas públicas (sem token):**
    - `POST /auth/login`
    - `POST /clientes`
    - `POST /bancos`
    - `POST /empresas`
    - `GET /swagger-ui`
    - `GET /swagger-ui/swagger.yml`
  version: 0.1.0
servers:
  - url: http://localhost:8080/api/aluguelcarros/v1
    description: Ambiente local
...
paths:
  /auth/login:
    post:
      tags: [Autenticação]
      summary: Autenticar e obter token JWT
      description: |
        Rota pública.

        O campo `login` aceita:
        - CPF para cliente
        - CNPJ para empresa ou banco
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              login: "12345678901"
              senha: "senha123"
      responses:
        '200':
          description: Autenticação bem-sucedida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensagemErro'
              example:
                mensagem: "Credenciais inválidas."
  /auth/me:
    get:
      tags: [Autenticação]
      summary: Retornar o perfil autenticado
      description: |
        Retorna dados básicos do usuário autenticado (`id`, `login`, `userType`).
      operationId: me
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
        '401':
          description: Token ausente, inválido ou formato incorreto
...
  /bancos:
    post:
      tags: [Bancos]
      summary: Cadastrar banco
      description: |
        Rota pública.

        O banco autentica por CNPJ e usa o CNPJ como login interno.
      operationId: cadastrarBanco
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BancoRequest'
            example:
              razaoSocial: "Banco XPTO"
              cnpj: "12345678000199"
              codigoBancario: "341"
              senha: "senhaBanco123"
      responses:
        '200':
          description: Banco criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BancoResponse'
        '400':
          description: Erro de validação
        '422':
          description: Regra de negócio violada
    get:
      tags: [Bancos]
      summary: Listar bancos
      operationId: listarBancos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de bancos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BancoResponse'
        '401':
          description: Não autorizado
  /bancos/{id}:
    parameters:
      - $ref: '#/components/parameters/BancoId'
    get:
      tags: [Bancos]
      summary: Buscar banco por id
      operationId: buscarBancoPorId
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Banco encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BancoResponse'
        '401':
          description: Não autorizado
        '404':
          description: Banco não encontrado
    put:
      tags: [Bancos]
      summary: Atualizar banco
      operationId: atualizarBanco
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BancoRequest'
      responses:
        '200':
          description: Banco atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BancoResponse'
        '400':
          description: Erro de validação
        '401':
          description: Não autorizado
        '404':
          description: Banco não encontrado
        '422':
          description: Regra de negócio violada
    delete:
      tags: [Bancos]
      summary: Remover banco
      operationId: removerBanco
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Banco removido
        '401':
          description: Não autorizado
        '404':
          description: Banco não encontrado
...
  /automoveis:
    post:
      tags: [Automóveis]
      summary: Cadastrar automóvel
      operationId: cadastrarAutomovel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutomovelRequest'
            example:
              matricula: "MAT-1001"
              placa: "ABC1D23"
              ano: 2024
              marca: "Toyota"
              modelo: "Corolla"
              status: "DISPONIVEL"
      responses:
        '200':
          description: Automóvel criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomovelResponse'
        '400':
          description: Erro de validação
        '401':
          description: Não autorizado
        '422':
          description: Regra de negócio violada
    get:
      tags: [Automóveis]
      summary: Listar automóveis
      operationId: listarAutomoveis
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de automóveis
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AutomovelResponse'
        '401':
          description: Não autorizado
  /automoveis/{id}:
    parameters:
      - $ref: '#/components/parameters/AutomovelId'
    get:
      tags: [Automóveis]
      summary: Buscar automóvel por id
      description: |
        Mesmo path base usado no backend para buscar por id (numérico)
        ou por status (texto). Quando o valor do path é numérico,
        o backend interpreta como id.
      operationId: buscarAutomovelPorId
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Automóvel encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomovelResponse'
        '401':
          description: Não autorizado
        '404':
          description: Automóvel não encontrado
    put:
      tags: [Automóveis]
      summary: Atualizar automóvel
      operationId: atualizarAutomovel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutomovelRequest'
      responses:
        '200':
          description: Automóvel atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomovelResponse'
        '400':
          description: Erro de validação
        '401':
          description: Não autorizado
        '404':
          description: Automóvel não encontrado
        '422':
          description: Regra de negócio violada
    delete:
      tags: [Automóveis]
      summary: Remover automóvel
      operationId: removerAutomovel
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Automóvel removido
        '401':
          description: Não autorizado
        '404':
          description: Automóvel não encontrado
  /automoveis/{status}:
    parameters:
      - $ref: '#/components/parameters/AutomovelStatusPath'
    get:
      tags: [Automóveis]
      summary: Listar automóveis por status
      description: |
        Endpoint de consulta por status solicitado.

        Status aceitos:
        - `DISPONIVEL`
        - `ALUGADO`
        - `VINCULADO`
      operationId: listarAutomoveisPorStatus
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de automóveis filtrada por status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AutomovelResponse'
        '401':
          description: Não autorizado
        '422':
          description: Status inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensagemErro'
...
  /agente/pedidos/em-analise:
    get:
      tags: [Fluxos de Negócio]
      summary: Listar pedidos em análise para agente
      description: Disponível para perfis `empresa` e `banco`.
      operationId: listarPedidosEmAnaliseParaAgente
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de pedidos com status EM_ANALISE
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PedidoResponse'
        '401':
          description: Não autorizado
        '422':
          description: Perfil sem permissão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensagemErro'
  /agente/pedidos/{id}/avaliar:
    parameters:
      - $ref: '#/components/parameters/PedidoId'
    post:
      tags: [Fluxos de Negócio]
      summary: Avaliar pedido como agente
      description: |
        Disponível para perfis `empresa` e `banco`.
        Cria parecer para o pedido e atualiza o status para `APROVADO` ou `REJEITADO`.
      operationId: avaliarPedidoComoAgente
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PedidoAvaliacaoRequest'
            example:
              resultado: "APROVADO"
              justificativa: "Renda e documentação aprovadas."
      responses:
        '200':
          description: Parecer emitido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParecerResponse'
        '400':
          description: Erro de validação
        '401':
          description: Não autorizado
        '404':
          description: Pedido não encontrado
        '422':
          description: Regra de negócio violada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensagemErro'
  /banco/pedidos/{id}/associar-credito:
    parameters:
      - $ref: '#/components/parameters/PedidoId'
    post:
      tags: [Fluxos de Negócio]
      summary: Associar crédito ao pedido
      description: |
        Disponível somente para perfil `banco`.
        Cria (ou reaproveita) contrato de tipo `BANCO` e associa contrato de crédito.
      operationId: associarCreditoAoPedido
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssociarCreditoRequest'
            example:
              valorFinanciado: 50000.0
              taxaJuros: 1.25
              prazoMeses: 36
              dataAssinatura: "2026-04-14"
      responses:
        '200':
          description: Crédito associado ao pedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContratoCreditoResponse'
        '400':
          description: Erro de validação
        '401':
          description: Não autorizado
        '404':
          description: Pedido/Contrato/Banco não encontrado
        '422':
          description: Regra de negócio violada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensagemErro'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token retornado em `POST /auth/login`
  parameters:
...
    BancoId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    AutomovelId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    AutomovelStatusPath:
      name: status
      in: path
      required: true
      schema:
        type: string
    PedidoId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
    LoginRequest:
      type: object
      required: [login, senha]
      properties:
        login:
          type: string
          description: CPF (cliente) ou CNPJ (empresa/banco)
        senha:
          type: string
          format: password
    TokenResponse:
      type: object
      required: [token, expiresIn, userType]
      properties:
        token:
          type: string
          description: JWT
        expiresIn:
          type: integer
          format: int64
          description: Tempo de expiração do token em milissegundos
        userType:
          $ref: '#/components/schemas/UserType'
    AuthMeResponse:
      type: object
      required: [id, login, userType]
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        userType:
          $ref: '#/components/schemas/UserType'
    UserType:
      type: string
      enum: [cliente, empresa, banco]
...
    BancoRequest:
      type: object
      required: [razaoSocial, cnpj, codigoBancario, senha]
      properties:
        razaoSocial:
          type: string
        cnpj:
          type: string
        codigoBancario:
          type: string
        senha:
          type: string
          format: password
    BancoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        razaoSocial:
          type: string
        cnpj:
          type: string
        codigoBancario:
          type: string
...
    AutomovelStatus:
      type: string
      enum: [DISPONIVEL, ALUGADO, VINCULADO]
    AutomovelRequest:
      type: object
      required: [matricula, placa, ano, marca, modelo, status]
      properties:
        matricula:
          type: string
        placa:
          type: string
        ano:
          type: integer
          format: int32
        marca:
          type: string
        modelo:
          type: string
        status:
          $ref: '#/components/schemas/AutomovelStatus'
    AutomovelResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        matricula:
          type: string
        placa:
          type: string
        ano:
          type: integer
          format: int32
        marca:
          type: string
        modelo:
          type: string
        status:
          $ref: '#/components/schemas/AutomovelStatus'
    PedidoStatus:
      type: string
      enum: [EM_ANALISE, APROVADO, REJEITADO, CANCELADO]
...
    PedidoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        numeroProtocolo:
          type: string
        clienteId:
          type: integer
          format: int64
        automovelId:
          type: integer
          format: int64
        dataInicio:
          type: string
          format: date
        dataFim:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/PedidoStatus'
    AgenteTipo:
      type: string
      enum: [BANCO, EMPRESA]
    ParecerResultado:
      type: string
      enum: [APROVADO, REJEITADO]
...
    ParecerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        pedidoId:
          type: integer
          format: int64
        agenteTipo:
          $ref: '#/components/schemas/AgenteTipo'
        agenteId:
          type: integer
          format: int64
        resultado:
          $ref: '#/components/schemas/ParecerResultado'
        justificativa:
          type: string
        dataEmissao:
          type: string
          format: date
...
    PedidoAvaliacaoRequest:
      type: object
      required: [resultado, justificativa]
      properties:
        resultado:
          $ref: '#/components/schemas/ParecerResultado'
        justificativa:
          type: string
    AssociarCreditoRequest:
      type: object
      required: [valorFinanciado, taxaJuros, prazoMeses]
      properties:
        valorFinanciado:
          type: number
          format: double
          minimum: 0.01
        taxaJuros:
          type: number
          format: double
          minimum: 0.01
        prazoMeses:
          type: integer
          format: int32
          minimum: 1
        dataAssinatura:
          type: string
          format: date
    ContratoCreditoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        contratoId:
          type: integer
          format: int64
        bancoId:
          type: integer
          format: int64
        valorFinanciado:
          type: number
          format: double
        taxaJuros:
          type: number
          format: double
        prazoMeses:
          type: integer
          format: int32
    MensagemErro:
      type: object
      properties:
        mensagem:
          type: string